% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{disconnect_cohort_connector}
\alias{disconnect_cohort_connector}
\title{Disconnect a persistent cohort_omop_connector}
\usage{
disconnect_cohort_connector(connector)
}
\arguments{
\item{connector}{A \code{cohort_connector} object.}
}
\value{
\code{connector} invisibly, with \verb{$conn} set to \code{NULL}.
}
\description{
Disconnect a persistent cohort_omop_connector
}
