% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_cohort_members}
\alias{extract_cohort_members}
\title{Extract cohort member records}
\usage{
extract_cohort_members(
  connector,
  cohort_definition_id = 1L,
  cohort_table = "cohort"
)
}
\arguments{
\item{connector}{A \code{cohort_connector} object.}

\item{cohort_definition_id}{\code{integer(1)}. Cohort definition ID to filter.}

\item{cohort_table}{\code{character(1)}. Name of the cohort table.}
}
\value{
tibble(subject_id, cohort_start_date, cohort_end_date)
}
\description{
Extract cohort member records
}
