% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_person_demographics}
\alias{extract_person_demographics}
\title{Extract person-level demographic information}
\usage{
extract_person_demographics(connector, subject_ids)
}
\arguments{
\item{connector}{A \code{cohort_connector} object.}

\item{subject_ids}{Integer vector of patient IDs.}
}
\value{
tibble with person demographics.
}
\description{
Extract person-level demographic information
}
