% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launch_cohort_intelligence}
\alias{launch_cohort_intelligence}
\title{Launch the CohortIntelligence Shiny dashboard}
\usage{
launch_cohort_intelligence(...)
}
\arguments{
\item{...}{Additional arguments forwarded to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\description{
Launch the CohortIntelligence Shiny dashboard
}
