% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{list_cohort_templates}
\alias{list_cohort_templates}
\title{List ATLAS JSON cohort templates bundled with CohortIntelligence}
\usage{
list_cohort_templates()
}
\value{
Named character vector mapping cohort names to file paths.
}
\description{
List ATLAS JSON cohort templates bundled with CohortIntelligence
}
